requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
cuda-bindings==12.9.0
cuda-python==12.9.0
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.31.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.7.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.15
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
